+++ src/Movie.cpp
@@ -21,7 +21,7 @@ Watchable* Movie::clone(){ return new Movie(*this);}
 std::string Movie::toString() const {
     std::vector<std::string> tag = getTag();
     std::string tagString = "[";
-    for (int i=0;i<tag.size();i++){
+    for (int i=0;(unsigned)i<tag.size();i++){
         if (i>0){
             tagString = tagString + ",";
         }
