+++ src/PrintWatchHistory.cpp
@@ -11,7 +11,7 @@
 void PrintWatchHistory::act(Session &sess) {
     std::vector<Watchable *> myhistory = sess.getActiveUser()->get_history();
 
-    for (int i = 0; i < myhistory.size(); i++) {
+    for (int i = 0; (unsigned) i < myhistory.size(); i++) {
         std::cout << myhistory[i]->toString() << std::endl;
     }
     complete();
